refactor(config): split merge.rs into focused submodules
Replaces the 1783-line merge.rs with a merge/ module containing 8
focused files, each under 500 lines, to improve token efficiency for
agents reading the codebase.

Module layout:
- types.rs     — ConfigLayer enum, FieldSources type alias
- path.rs      — get_nested, set_nested, remove_nested helpers
- deep.rs      — deep_merge, deep_merge_tracking, record_all_leaves
- enforce.rs   — 8 enforcement helper functions (clamp, bool, union, etc.)
- servers.rs   — PROTECTED_SERVER_FIELDS, sanitize_workspace_servers
- restrict.rs  — enforce_restrictions orchestrator
- tests.rs     — all 27 tests (consolidated)

No logic changes; all 93 tests pass.
